chore: Update .env and app.ts files
The changes in this commit update the `.env` file to add new environment variables for `AGENT`, `VERSION`, `REGION`, `SUBREGION`, `SERVER`, and `EDGE`. It also updates the `app.ts` file to include the `headersMiddleware` for setting headers in the Express app. These changes improve the configuration and functionality of the application.