feat: Include match duration in GetLastMatchesService
This commit adds the match duration to the response of the GetLastMatchesService. The match duration is obtained from the match data and is included in both milliseconds and minutes format in the `matchDurationMilis` and `matchDurationMinutes` fields of the match object. This enhancement provides additional information about the match and improves the functionality of the application.