refactor: Update lastmatches.js to use correct data path for map name and image
This commit updates the lastmatches.js file to use the correct data path for the map name and image. Previously, the incorrect path was being used, resulting in the map name and image not being displayed correctly. By fixing the data path, the map name and image will now be displayed correctly in the application.