refactor: Update fetch URL for lastmatches.js to use correct endpoint
This commit updates the fetch URL in the lastmatches.js file to use the correct endpoint '/v1/riot/player/last-matches'. Previously, the incorrect endpoint '/v1/riot/dev/player/last-matches' was being used, causing the data for map name not to be displayed correctly. By fixing the fetch URL, the map name will now be displayed accurately in the application.